# workspace/fairsplit/api-spec.yaml
openapi: 3.1.0
info:
  title: FairSplit API
  version: 1.0.0
  description: |
    Zero-dependency shared expense tracker API.

    ## Authentication
    All group-scoped endpoints require a valid member session token.
    The token is issued by `POST /api/v1/groups` (group creation) or
    `POST /api/v1/groups/{group_id}/members` (join via invite link).

    **Primary:** `httpOnly` cookie named `fairsplit_member_token`.
    **Fallback:** `Authorization: Bearer <token>` header (API clients).

    The token is a JWT signed with HS256 containing:
    ```json
    {
      "sub": "<member_id>",
      "group_id": "<group_id>",
      "is_admin": true,
      "iat": 1713200000,
      "exp": 1715792000
    }
    ```

    ## Response Envelopes
    **Singleton responses** (auth, single-resource GET/POST/PATCH/DELETE):
    The resource object is returned directly at the top level. No wrapper.

    **List responses** (collection GET endpoints):
    ```json
    {
      "data": [...],
      "total": 23,
      "page": 1,
      "per_page": 20
    }
    ```

    **Error responses** (all 4xx and 5xx):
    ```json
    {
      "error": {
        "code": "VALIDATION_ERROR",
        "message": "Human-readable description.",
        "field": "amount",
        "request_id": "req_abc123"
      }
    }
    ```

    ## Idempotency
    Mutation endpoints (`POST` for create, `DELETE` for soft-delete) accept an
    `X-Idempotency-Key` header containing a client-generated UUID. The server
...

  contact:
...
servers:
...
paths:
...
  /groups/{group_id}/balances:
    get:
      operationId: getBalances
      summary: Get net balances for all members
      description: |
        Computes and returns the net balance for every member in the group.
        Net balance = total paid - total owed from expense splits + settlements paid - settlements received.
        All amounts are in integer cents.

        This endpoint is called by the frontend every 10 seconds (SWR polling).
        It must complete in < 200ms at p95 for groups up to 50 members and
        200 expenses.

        Positive `net_cents` = creditor (owed money, shown in green).
        Negative `net_cents` = debtor (owes money, shown in red).
        Zero `net_cents` = settled (shown in gray).
      tags: [balances]
      security:
        - MemberCookie: []
        - BearerToken: []
      parameters:
        - $ref: "#/components/parameters/GroupId"
      responses:
        "200":
          description: Net balances for all members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalancesResponse"
              example:
                currency: USD
                members:
                  - member_id: "a3bb1891-1f4e-4e7a-8c4f-6d4a1e2b3c4d"
                    display_name: "Maya"
                    net_cents: 8450
                    net_display: "+$84.50"
                    paid_total_cents: 24000
                    owed_total_cents: 15550
                    settled_out_cents: 0
                    settled_in_cents: 0
                    status: creditor
                  - member_id: "b4cc2902-2g5f-5f8b-9d5g-7e5b2f3c4d5e"
                    display_name: "Daniel"
                    net_cents: -4200
                    net_display: "-$42.00"
                    paid_total_cents: 0
                    owed_total_cents: 4200
                    settled_out_cents: 0
                    settled_in_cents: 0
                    status: debtor
                computed_at: "2026-04-15T18:30:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /groups/{group_id}/balances/{member_id}:
    get:
      operationId: getMemberBalance
      summary: Get detailed balance breakdown for a single member
      description: |
        Returns the balance breakdown for a single member: total paid,
        total owed, settlements, and net balance. Used in the "tap a member
        balance" detail view.
      tags: [balances]
      security:
        - MemberCookie: []
        - BearerToken: []
      parameters:
        - $ref: "#/components/parameters/GroupId"
        - name: member_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The member's UUID
      responses:
        "200":
          description: Member balance detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberBalanceDetail"
              example:
                member_id: "a3bb1891-1f4e-4e7a-8c4f-6d4a1e2b3c4d"
                display_name: "Maya"
                net_cents: 8450
                net_display: "+$84.50"
                paid_total_cents: 24000
                owed_total_cents: 15550
                settled_out_cents: 0
                settled_in_cents: 0
                expense_count_paid: 8
                summary: "Maya paid $240.00 across 8 expenses. Her share was $155.50. Net: +$84.50."
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ============================================================
  # SETTLE-UP
  # ============================================================
...
components:
  securitySchemes:
    MemberCookie:
      type: apiKey
      in: cookie
      name: fairsplit_member_token
      description: |
        httpOnly session cookie issued on group creation or group join.
        Set by the server. Cannot be accessed by JavaScript.
        Expiry: 30 days, rolling (reset on each valid request).
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Fallback for API clients that cannot use cookies.
        Same JWT as the cookie value. Pass in Authorization: Bearer <token> header.
  parameters:
    GroupId:
      name: group_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: The group's UUID
...
  responses:
    Unauthorized:
      description: No session or session expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            no_session:
              value:
                error:
                  code: NO_SESSION
                  message: "No session token found. Please join the group first."
                  request_id: req_abc123
            session_expired:
              value:
                error:
                  code: SESSION_EXPIRED
                  message: "Your session has expired. Please re-enter your name to continue."
                  request_id: req_abc123
    Forbidden:
      description: Valid session but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            permission_denied:
              value:
                error:
                  code: PERMISSION_DENIED
                  message: "You can only edit expenses you logged, or you must be the group admin."
                  request_id: req_abc123
            wrong_group:
              value:
                error:
                  code: MEMBER_WRONG_GROUP
                  message: "Your session token is for a different group."
                  request_id: req_abc123
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: GROUP_NOT_FOUND
              message: "Group not found. The link may be invalid."
              request_id: req_abc123
...
  schemas:
...
    MemberBalance:
      type: object
      required:
        - member_id
        - display_name
        - net_cents
        - net_display
        - paid_total_cents
        - owed_total_cents
        - settled_out_cents
        - settled_in_cents
        - status
      properties:
        member_id:
          type: string
          format: uuid
        display_name:
          type: string
        net_cents:
          type: integer
          description: Net balance in integer cents. Positive = creditor, negative = debtor
        net_display:
          type: string
          description: Formatted display string (e.g., "+$84.50" or "-$42.00")
        paid_total_cents:
          type: integer
          minimum: 0
          description: Total amount this member has paid (as payer) across all active expenses
        owed_total_cents:
          type: integer
          minimum: 0
          description: Total amount this member owes (from expense_splits)
        settled_out_cents:
          type: integer
          minimum: 0
          description: Total amount this member has paid in settlements
        settled_in_cents:
          type: integer
          minimum: 0
          description: Total amount this member has received in settlements
        status:
          type: string
          enum: [creditor, debtor, settled]
    MemberBalanceDetail:
      allOf:
        - $ref: "#/components/schemas/MemberBalance"
        - type: object
          required: [expense_count_paid, summary]
          properties:
            expense_count_paid:
              type: integer
              description: Number of expenses where this member was the payer
            summary:
              type: string
              description: |
                Human-readable one-sentence balance summary.
                Example: "Maya paid $240.00 across 8 expenses. Her share was $155.50. Net: +$84.50."
    BalancesResponse:
      type: object
      required: [currency, members, computed_at]
      properties:
        currency:
          type: string
          pattern: "^[A-Z]{3}$"
          description: The currency for which balances are computed
        members:
          type: array
          items:
            $ref: "#/components/schemas/MemberBalance"
          description: All members sorted by net_cents DESC (largest creditor first)
        computed_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - GROUP_NOT_FOUND
                - MEMBER_NOT_FOUND
                - EXPENSE_NOT_FOUND
                - SETTLEMENT_NOT_FOUND
                - INVALID_INVITE_TOKEN
                - GROUP_ARCHIVED
                - GROUP_HAS_OUTSTANDING_BALANCES
                - PERMISSION_DENIED
                - NO_SESSION
                - SESSION_EXPIRED
                - SESSION_INVALID
                - MEMBER_WRONG_GROUP
                - IDEMPOTENCY_CONFLICT
                - SPLIT_SUM_MISMATCH
                - RATE_LIMIT_EXCEEDED
                - INTERNAL_ERROR
            message:
              type: string
              description: Human-readable error description
            field:
              type: string
              nullable: true
              description: The specific field that failed validation (for VALIDATION_ERROR only)
            request_id:
              type: string
              description: Unique request identifier for log correlation